# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Imposter AI API
  version: 1.0.0
  description: |
    # Imposter AI — Backend Specification

    Backend API для игры Imposter AI — управление категориями, AI-генерацией слов, premium-контентом и remote config.

    ---

    ## Сущности

    ### Category
    Описывает тематическую категорию слов. Делятся на free и premium через флаг `is_premium`.
    Активность управляется через `is_active` — скрывает без удаления. `sort_order` определяет порядок в клиенте.
    Все текстовые поля хранятся как LocalizedString — JSON-объект с ключами по ISO locale.

    ### WordPack
    Набор слов привязанный к категории. Создаётся отдельно под каждую локаль — автоматического перевода нет.
    - Для стандартного режима `impostor_word = null` — impostor не получает слово.
    - Для Party Mode (Undercover) `impostor_word` содержит похожее но отличное слово.
    - При запросе клиент передаёт `locale`, сервер ищет подходящий пак; если нет — fallback на `en`.

    ### FeatureConfig
    Управление feature flags и контент-гейтингом. Клиент получает весь конфиг одним запросом `GET /config` при старте.
    Флаги кешируются на клиенте с TTL 15–30 мин. `payload` может содержать лимиты AI-запросов для free-tier.
    Ключи: `"ai_theme"`, `"premium_categories"`, `"party_mode_extra_roles"`.

    ### AITopicRequestLog
    Лог запросов на AI-генерацию — для дебага, модерации, rate limiting. Хранится для аудита, не возвращается клиенту.
...

    ### LocalizationContent
...

    ---

...

...

...

...

...

...

...

...
servers:
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
  parameters:
    DeviceId:
      name: X-Device-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Анонимный идентификатор устройства, генерируется при первом запуске
    AppVersion:
      name: X-App-Version
      in: header
      required: true
      schema:
        type: string
        example: "1.0.0"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: RATE_LIMIT_EXCEEDED
            message:
              type: string
              example: AI request limit reached for today
            details:
              nullable: true
              example: null
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Внутренний идентификатор пользователя
        device_id:
          type: string
          format: uuid
          description: Анонимный идентификатор устройства
        tokens:
          type: integer
          default: 0
          description: Текущий баланс токенов
        is_premium:
          type: boolean
          description: Наличие активной premium-подписки
        premium_expires_at:
          type: string
          format: date-time
          nullable: true
          description: Дата истечения premium-подписки (null если нет подписки)
        created_at:
          type: string
          format: date-time
          description: Дата и время регистрации пользователя
      required:
        - id
        - device_id
        - tokens
        - is_premium
        - created_at
    CreateUserRequest:
      type: object
      required: [device_id]
      properties:
        device_id:
          type: string
          format: uuid
          description: Анонимный идентификатор устройства
...
paths:
...
  /users:
    post:
      summary: Создание пользователя
      description: |
        Регистрирует нового пользователя по device_id и сразу возвращает профиль + JWT access_token.
        Если пользователь с таким device_id уже существует — возвращает существующий профиль (idempotent).
        Не требует X-Api-Key — вызывается при первом запуске до получения токена.
      tags: [User]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: Профиль существующего пользователя (upsert — пользователь уже был создан ранее)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UserProfile'
                  - type: object
                    required: [access_token, token_type]
                    properties:
                      access_token:
                        type: string
                        description: JWT Bearer токен
                        example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                      token_type:
                        type: string
                        enum: [bearer]
                        example: bearer
        '201':
          description: Профиль нового пользователя
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UserProfile'
                  - type: object
                    required: [access_token, token_type]
                    properties:
                      access_token:
                        type: string
                        description: JWT Bearer токен
                        example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                      token_type:
                        type: string
                        enum: [bearer]
                        example: bearer
              example:
                id: "550e8400-e29b-41d4-a716-446655440100"
                device_id: "550e8400-e29b-41d4-a716-446655440099"
                tokens: 0
                is_premium: false
                premium_expires_at: null
                created_at: "2026-06-09T10:00:00Z"
                access_token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                token_type: bearer
        '400':
          description: Невалидный device_id (не UUID)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            Возвращается если передан явный конфликт. В стандартном сценарии (upsert по device_id) операция idempotent — 409 не возвращается, вместо этого приходит 200 с существующим профилем.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/me:
    get:
      summary: Профиль текущего пользователя
      description: |
        Возвращает профиль пользователя идентифицированного по X-Device-Id.
        Требует стандартный X-Api-Key + X-Device-Id header.
        Идентификация пользователя происходит через X-Device-Id.
      tags: [User]
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/AppVersion'
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
              example:
                id: "550e8400-e29b-41d4-a716-446655440100"
                device_id: "550e8400-e29b-41d4-a716-446655440099"
                tokens: 15
                is_premium: true
                premium_expires_at: "2026-07-09T10:00:00Z"
                created_at: "2026-06-09T10:00:00Z"
        '401':
          description: Отсутствует или невалидный X-Api-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Пользователь с таким device_id не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...